+++ wpilibc/sim/include/AnalogInput.h
@@ -31,17 +31,17 @@ class AnalogInput : public SensorBase,
                     public PIDSource,
                     public LiveWindowSendable {
  public:
-  static const uint8_t kAccumulatorModuleNumber = 1;
-  static const uint32_t kAccumulatorNumChannels = 2;
-  static const uint32_t kAccumulatorChannels[kAccumulatorNumChannels];
+  static const int kAccumulatorModuleNumber = 1;
+  static const int kAccumulatorNumChannels = 2;
+  static const int kAccumulatorChannels[kAccumulatorNumChannels];
 
-  explicit AnalogInput(uint32_t channel);
+  explicit AnalogInput(int channel);
   virtual ~AnalogInput() = default;
 
   float GetVoltage() const;
   float GetAverageVoltage() const;
 
-  uint32_t GetChannel() const;
+  int GetChannel() const;
 
   double PIDGet() override;
 
@@ -53,7 +53,7 @@ class AnalogInput : public SensorBase,
   std::shared_ptr<ITable> GetTable() const override;
 
  private:
-  uint32_t m_channel;
+  int m_channel;
   SimFloatInput* m_impl;
   int64_t m_accumulatorOffset;
